Main Menu Screen made for VR
Main menu screen is now a desk with a computer on it. when the player looks at the computer a tooltip apears. when they click on the computer the game loads the next scene.  also i fixed the bug that caused the player to start out looking in the wrong direction at the begining of the scene.